+++ src/LSST/M1M3/SS/Accelerometer/AccelerometerSettings.h
@@ -46,6 +46,12 @@ public:
     void load(YAML::Node doc);
 
     void log() { M1M3SSPublisher::instance().logAccelerometerSettings(this); }
+
+    double xElevationPoly[3];
+    double yElevationPoly[3];
+    double zElevationPoly[3];
+
+    std::string dump_path;
 };
 
 } /* namespace SS */