+++ src/LSST/M1M3/SS/Controllers/ForceController.cpp
@@ -211,8 +211,15 @@ void ForceController::updateAppliedForces() {
             _balanceForceComponent.applyBalanceForcesByMirrorForces(
                     _hardpointActuatorData->fx, _hardpointActuatorData->fy, _hardpointActuatorData->fz,
                     _hardpointActuatorData->mx, _hardpointActuatorData->my, _hardpointActuatorData->mz);
+        } else {
+            _balanceForceComponent.applyFreezedForces();
         }
         _balanceForceComponent.update();
+    } else {
+        bool changed = _balanceForceComponent.applyFreezedForces();
+        if (changed) {
+            _balanceForceComponent.update();
+        }
     }
     if (_elevationForceComponent.isActive()) {
         if (_elevationForceComponent.isEnabled()) {
@@ -341,6 +348,16 @@ void ForceController::resetPID(int id) {
 void ForceController::resetPIDs() {
     SPDLOG_INFO("ForceController: resetPIDs()");
     _balanceForceComponent.resetPIDs();
+}
+
+void ForceController::freezePIDs() {
+    SPDLOG_INFO("ForceController: freezePIDs()");
+    _balanceForceComponent.freezePIDs();
+}
+
+void ForceController::thawPIDs() {
+    SPDLOG_INFO("ForceController: thawPIDs()");
+    _balanceForceComponent.thawPIDs();
 }
 
 void ForceController::applyElevationForces() {