+++ src/LSST/M1M3/SS/ForceComponents/BalanceForceComponent.cpp
@@ -54,10 +54,10 @@ BalanceForceComponent::BalanceForceComponent(
     _preclippedBalanceForces = M1M3SSPublisher::instance().getEventPreclippedBalanceForces();
 }
 
-void BalanceForceComponent::applyBalanceForces(float* x, float* y, float* z) {
+void BalanceForceComponent::applyBalanceForces(float* x, float* y, float* z, bool check) {
     SPDLOG_TRACE("BalanceForceComponent: applyBalanceForces()");
 
-    if (!isEnabled()) {
+    if (check && !isEnabled()) {
         SPDLOG_ERROR("BalanceForceComponent: applyBalanceForces() called when the component is not applied");
         return;
     }
@@ -109,6 +109,35 @@ void BalanceForceComponent::applyBalanceForcesByMirrorForces(float xForce, float
     applyBalanceForces(xForces, yForces, zForces);
 }
 
+bool BalanceForceComponent::applyFreezedForces() {
+    bool changed = false;
+    float fx = _fx.getOffset(&changed);
+    float fy = _fy.getOffset(&changed);
+    float fz = _fz.getOffset(&changed);
+    float mx = _mx.getOffset(&changed);
+    float my = _my.getOffset(&changed);
+    float mz = _mz.getOffset(&changed);
+    DistributedForces forces =
+            ForceActuatorSettings::instance().calculateForceDistribution(fx, fy, fz, mx, my, mz);
+    float xForces[FA_X_COUNT];
+    float yForces[FA_Y_COUNT];
+    float zForces[FA_Z_COUNT];
+    for (int zIndex = 0; zIndex < FA_COUNT; ++zIndex) {
+        int xIndex = _forceActuatorApplicationSettings->ZIndexToXIndex[zIndex];
+        int yIndex = _forceActuatorApplicationSettings->ZIndexToYIndex[zIndex];
+
+        if (xIndex != -1) {
+            xForces[xIndex] = forces.XForces[zIndex];
+        }
+        if (yIndex != -1) {
+            yForces[yIndex] = forces.YForces[zIndex];
+        }
+        zForces[zIndex] = forces.ZForces[zIndex];
+    }
+    applyBalanceForces(xForces, yForces, zForces, false);
+    return changed;
+}
+
 void BalanceForceComponent::updatePID(int id, PIDParameters parameters) {
     SPDLOG_DEBUG("BalanceForceComponent: updatePID({})", id);
     PID* pid = _idToPID(id);
@@ -133,6 +162,26 @@ void BalanceForceComponent::resetPIDs() {
     _mx.restoreInitialParameters();
     _my.restoreInitialParameters();
     _mz.restoreInitialParameters();
+}
+
+void BalanceForceComponent::freezePIDs() {
+    SPDLOG_INFO("BalanceForceComponent: freezePIDs()");
+    _fx.freeze();
+    _fy.freeze();
+    _fz.freeze();
+    _mx.freeze();
+    _my.freeze();
+    _mz.freeze();
+}
+
+void BalanceForceComponent::thawPIDs() {
+    SPDLOG_INFO("BalanceForceComponent: thawPIDs()");
+    _fx.thaw();
+    _fy.thaw();
+    _fz.thaw();
+    _mx.thaw();
+    _my.thaw();
+    _mz.thaw();
 }
 
 void BalanceForceComponent::postEnableDisableActions() {