+++ src/LSST/M1M3/SS/ForceComponents/BalanceForceComponent.h
@@ -55,7 +55,7 @@ class BalanceForceComponent : public ForceComponent {
 public:
     BalanceForceComponent(ForceActuatorApplicationSettings* forceActuatorApplicationSettings);
 
-    void applyBalanceForces(float* x, float* y, float* z);
+    void applyBalanceForces(float* x, float* y, float* z, bool check = true);
 
     /**
      * Called from ForceController::updateAppliedForces. Feeds in hardpoint
@@ -71,10 +71,14 @@ public:
      */
     void applyBalanceForcesByMirrorForces(float xForce, float yForce, float zForce, float xMoment,
                                           float yMoment, float zMoment);
+    bool applyFreezedForces();
 
     void updatePID(int id, PIDParameters parameters);
     void resetPID(int id);
     void resetPIDs();
+
+    void freezePIDs();
+    void thawPIDs();
 
     void postEnableDisableActions() override;