+++ doc/version-history.rst
@@ -7,6 +7,7 @@ Version History
 v2.15.2
 -------
 
+* utils dcelevation, dcrawdump utilities
 * Read and store raw DC accelerometer data
 
 v2.15.1
@@ -16,6 +17,7 @@ v2.15.1
 * Use cRIO/SAL/Command.h ACK_ constants
 * ReloadConfigurationCommand, fix ForceComponent's settings reloading
 * Fix DC accelerometer unit - it's in degrees, needs to be feed as radians to acceleration force calculation
+* Fix X DC accelerometer angular acceleration sign
 * Fixed Makefiles
 
 v2.15.0